requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pygame==2.1.2
pygame-gui==0.6.8
python-dotenv==0.21.1
python-i18n==0.3.9
requests==2.28.2
urllib3==1.26.14
